refactor hook hierarchy; add handling for keypoint-moved event
